refactor: flatten locale folders to drop en directory naming
