+++ docs/uml-diagrams/diagrams/png/activity-diagram-simplified.png
