+++ docs/uml-diagrams/diagrams/png/activity-diagram.png
