+++ docs/uml-diagrams/diagrams/png/class-diagram-patterns.png
