+++ docs/uml-diagrams/diagrams/png/class-diagram.png
