+++ docs/uml-diagrams/diagrams/png/sequence-diagram.png
