+++ docs/uml-diagrams/diagrams/puml/activity-diagram-simplified.puml
