+++ docs/uml-diagrams/diagrams/puml/class-diagram-patterns.puml
