+++ docs/uml-diagrams/diagrams/puml/class-diagram.puml
