+++ examples/analyze_pendlers_case.py
@@ -1,6 +1,85 @@
-"""Convenience wrapper that delegates to the English Pendlers case analysis."""
+"""
+Detailed BeCoMe analysis of Pendlers case study.
 
-from examples.en.analyze_pendlers_case import main
+Demonstrates step-by-step calculation for 22 expert Likert scale ratings
+(0-25-50-75-100) for cross-border travel policy.
+"""
+
+from pathlib import Path
+
+from examples.utils.analysis import calculate_agreement_level
+from examples.utils.data_loading import load_data_from_txt
+from examples.utils.display import (
+    display_step_1_arithmetic_mean,
+    display_step_2_median,
+    display_step_3_best_compromise,
+    display_step_4_max_error,
+)
+from examples.utils.formatting import display_case_header, print_header, print_section
+from examples.utils.locales import EN_ANALYSIS, EN_DISPLAY, EN_FORMATTING
+from src.calculators.base_calculator import BaseAggregationCalculator
+from src.calculators.become_calculator import BeCoMeCalculator
+from src.interpreters.likert_interpreter import LikertDecisionInterpreter
+
+
+def main(calculator: BaseAggregationCalculator | None = None) -> None:
+    """
+    Run detailed BeCoMe analysis for Pendlers case study.
+
+    :param calculator: Aggregation calculator instance.
+                       Defaults to BeCoMeCalculator if not provided.
+    """
+    if calculator is None:
+        calculator = BeCoMeCalculator()
+
+    data_file = str(Path(__file__).parent / "data" / "pendlers_case.txt")
+    opinions, metadata = load_data_from_txt(data_file)
+
+    display_case_header("PENDLERS CASE", opinions, metadata, EN_FORMATTING)
+
+    print("\nLikert scale interpretation:")
+    print("  0   = Strongly disagree")
+    print("  25  = Rather disagree")
+    print("  50  = Neutral")
+    print("  75  = Rather agree")
+    print("  100 = Strongly agree")
+
+    print("\nNote: For Likert scale, lower = peak = upper (crisp values)")
+    mean, mean_centroid = display_step_1_arithmetic_mean(opinions, calculator, EN_DISPLAY)
+    print(f"Mean centroid: {mean_centroid:.2f} (same as peak for crisp values)")
+
+    median, median_centroid = display_step_2_median(
+        opinions, calculator, is_likert=True, labels=EN_DISPLAY
+    )
+
+    best_compromise, bc_centroid = display_step_3_best_compromise(mean, median, EN_DISPLAY)
+
+    max_error = display_step_4_max_error(mean_centroid, median_centroid, EN_DISPLAY)
+
+    print_section("FINAL RESULT")
+    result = calculator.calculate_compromise(opinions)
+    print(f"\n{result}")
+
+    print_header("INTERPRETATION")
+
+    print(f"\nBest compromise estimate: {bc_centroid:.2f} (centroid)")
+    print(
+        f"Fuzzy number: ({best_compromise.lower_bound:.2f}, "
+        f"{best_compromise.peak:.2f}, {best_compromise.upper_bound:.2f})"
+    )
+    print(f"Precision indicator (Δmax): {max_error:.2f}")
+
+    agreement = calculate_agreement_level(max_error, thresholds=(5.0, 10.0), labels=EN_ANALYSIS)
+    print(f"Expert agreement: {agreement.upper()}")
+
+    interpreter = LikertDecisionInterpreter()
+    decision = interpreter.interpret(best_compromise)
+
+    print(f"Closest Likert value: {decision.likert_value}")
+    print(f"\nDECISION (based on centroid {bc_centroid:.2f}):")
+    print(f"  {decision.decision_text.upper()}")
+    print(f"\nRecommendation: {decision.recommendation}")
+
 
 if __name__ == "__main__":
     main()
