--- examples/data/en/budget_case.txt
@@ -1,28 +1,0 @@
-CASE: Budget
-DESCRIPTION: Experts express standpoint regarding state budget support and total financial support in the range CZK 0-100 billion for entrepreneurs affected by COVID-19 pandemic
-EXPERTS: 22
-
-# Format: ExpertID | Lower | Peak | Upper
-Chairman | 40 | 70 | 90
-Deputy chairman | 60 | 80 | 90
-Deputy Minister of MI | 60 | 90 | 90
-Deputy Minister of MD | 30 | 40 | 40
-Deputy Minister of MFA | 40 | 70 | 90
-Deputy Minister of MF | 30 | 45 | 70
-Deputy Minister of MH | 40 | 60 | 85
-Deputy Minister of MIT | 40 | 90 | 90
-Deputy Minister of MT | 40 | 50 | 50
-Deputy Minister of MEYS | 15 | 40 | 60
-Deputy Minister of MA | 40 | 50 | 70
-Deputy Minister of MoLSA | 10 | 50 | 50
-Deputy Minister of MC | 10 | 30 | 40
-Deputy Minister of MJ | 10 | 40 | 50
-Chairman of the SSHR | 40 | 40 | 40
-Police President | 50 | 50 | 50
-Director of Fire Rescue Service | 30 | 40 | 50
-Chief of General Staff of the ACR | 10 | 40 | 45
-Director of NÚKIB | 10 | 40 | 80
-Chief hygienist | 10 | 30 | 80
-Director of SZÚ | 30 | 50 | 55
-Director of the Office of the Government | 60 | 60 | 90
-
